requirements.txt
django 
SQLAlchemy
Pillow
opencv-python
django-phonenumber-field[phonenumberslite]
django-environ
django-crispy-forms
psycopg2